updates based on Torah comments
