+++ scripts/cognito-csv-exporter.js
@@ -2,54 +2,30 @@ const createCsvWriter = require('csv-writer').createArrayCsvWriter;
 const { userToCsvHeader } = require('./csv-utils/userToCsv');
 const AWS = require('aws-sdk');
 const globalAWSConfig = require('../globalAWSConfig');
+const getAllUsers = require('./cognito-utils/getAllUsers');
 
 // Configure the AWS SDK with your credentials
 AWS.config.update(globalAWSConfig);
 
 // Create an instance of the AWS Cognito Identity provider client
 const cognitoIdentityServiceProvider = new AWS.CognitoIdentityServiceProvider();
-const userPoolId = 'user-pool-id';
 
-// Cognito paginates users to return 60. We must loop through paginated results to grab all users
-const getAllUsers = async (params) => {
-  try {
-    // string must not be empty
-    let paginationToken = 'notEmpty';
-    let itemsAll = {
-      Users: [],
-    };
-    while (paginationToken) {
-      const data = await cognitoIdentityServiceProvider
-        .listUsers(params)
-        .promise();
+// @TODO: Update these to user pool values
+const USER_POOL_ID_CONTAINING_USERS_TO_MIGRATE = 'user-pool-id';
 
-      const { Users } = data;
-      itemsAll = {
-        ...data,
-        ...{ Users: [...itemsAll.Users, ...(Users ? [...Users] : [])] },
-      };
-      paginationToken = data.PaginationToken;
-      if (paginationToken) {
-        params.PaginationToken = paginationToken;
-      }
-    }
-    return itemsAll;
-  } catch (err) {
-    console.error('Unable to scan the cognito pool users. Error JSON:', err);
-  }
-};
+// provide the user pool ID of the new pool you want to transfer users to
+const USER_POOL_ID_FOR_CSV_HEADERS = 'user-pool-id';
 
 const init = async () => {
   const params = {
-    UserPoolId: userPoolId,
+    UserPoolId: USER_POOL_ID_CONTAINING_USERS_TO_MIGRATE,
   };
 
   const allUsers = await getAllUsers(params);
 
-  // provide the user pool ID of the new pool you want to transfer users to
   cognitoIdentityServiceProvider.getCSVHeader(
     {
-      UserPoolId: 'example-user-pool-id',
+      UserPoolId: USER_POOL_ID_FOR_CSV_HEADERS,
     },
     (err, csvData) => {
       if (err) console.error('oh no! an error occurred', err);
@@ -66,7 +42,8 @@ const init = async () => {
           .writeRecords(results) // returns a promise
           .then(() => {
             console.log('...Done');
-          });
+          })
+          .catch((err) => console.log('Issue writing records', err));
       }
     }
   );
